+++ R/utils-head.R
@@ -19,6 +19,10 @@
 #  tail,timeSeries           Returns the tail of a 'timeSeries' object
 ################################################################################
 
+setGeneric("head")
+setGeneric("tail")
+
+
 
 head.timeSeries <- 
     function(x, n = 6, recordIDs = FALSE, ...)
@@ -48,8 +52,6 @@ head.timeSeries <-
     else
         head.matrix(x, n = n, ...)
 }
-
-setMethod("head", "timeSeries", head.timeSeries)
 
           
 # ------------------------------------------------------------------------------
@@ -89,7 +91,8 @@ tail.timeSeries <-
         }
     }
 
-setMethod("tail", "timeSeries", tail.timeSeries)
-
+## (2024-01-05) GNB: stop making head() and tail() S4
+## setMethod("head", "timeSeries", head.timeSeries)
+## setMethod("tail", "timeSeries", tail.timeSeries)
 
 ################################################################################