+++ NEWS.md
@@ -1,3 +1,17 @@
+## timeSeries 4032.109
+
+- deprecated function `returnSeries` is now defunct, use `returns` instead.
+
+- a number of generic functions from base R now get only S3 methods for
+  'timeSeries' objects. Previously they were turned into S4 generics with S4
+  methods.
+
+- streamlined timeSeries methods for a number of functions. Left only S3 methods
+  or only S4 methods were suitable.
+
+- consolidated the NAMESPACE.
+
+
 ## timeSeries 4032.108
 
 - fixed 'Lost braces; missing escapes or markup?' NOTE from CRAN.